+++ src/datamatic/Generator.py
@@ -4,7 +4,7 @@ from typing import Tuple, Literal
 from dataclasses import dataclass
 from functools import partial
 
-from Plugins import Plugin
+from plugins import Plugin
 
 
 TOKEN = re.compile(r"\{\{(.*?)\}\}")
