+++ src/datamatic/Validator.py
@@ -2,7 +2,7 @@
 Validates a given schema to make sure it is well-formed. This should
 also serve as documentation for what makes a valid schema.
 """
-import Plugins
+import plugins
 
 
 COMP_KEYS_REQ = {
@@ -48,7 +48,7 @@ def validate_attribute(attr, flags):
     assert isinstance(attr["display_name"], str), attr
 
     # Verify that accessing the default value succeeds.
-    Plugins.builtin.default(attr)
+    plugins.builtin.default(attr)
 
     if "flags" in attr:
         assert isinstance(attr["flags"], dict)
